Added clunky version of list_for_parallelization
There are a lot of problems with this version of the algorithm to create a long list for the parallel code, but it returns a useable output. The function requires the first set of parameters to be listed as the second argument in a "column vector"-like list object the rest of the parameter ranges may be input normally in the first argument. An example of usage is given in the comment above the function definition.